refactor: extract power meter settings
