+++ utils/measure/frontend/src/components/settings/view.ts
@@ -1,16 +1,13 @@
 import { LitElement, css, html, nothing, svg, type PropertyValues } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 import { createRef, ref } from "lit/directives/ref.js";
-import type { AppSettings, AppSettingsUpdate, Capabilities, ContributionAuthDeviceStatus, ContributionAuthState, ContributionDeviceFlow, EntityDescriptor, MeasureParameterName, PowerMeterDiagnostic, PowerMeterType, SettingsSection, ShellyDiscoveryDevice } from "../../types";
-import { DEFAULT_SHELLY_USERNAME, POWER_METER_LIST, meterFor, settingsFromForm } from "../../power-meter";
+import type { AppSettings, AppSettingsUpdate, Capabilities, ContributionAuthDeviceStatus, ContributionAuthState, ContributionDeviceFlow, EntityDescriptor, MeasureParameterName, PowerMeterDiagnostic, SettingsSection, ShellyDiscoveryDevice } from "../../types";
+import { DEFAULT_SHELLY_USERNAME, settingsFromForm } from "../../power-meter";
 import { formRaw, formText, formTextOrNull } from "../../form";
 import { emit } from "../../events";
 import { sharedStyles } from "../../styles";
-import type { ComboboxOption } from "../shared/combobox";
-import "../shared/combobox";
-import { optionSelect } from "../shared/fields";
-import "../shared/power-meter-diagnostic";
 import "./github-section";
+import "./power-meter-section";
 
 interface SettingsSectionDescriptor {
   id: SettingsSection;
@@ -68,9 +65,6 @@ export class SettingsView extends LitElement {
   @property({ type: String })
   measureDevicesError = "";
 
-  @state()
-  meter?: PowerMeterType;
-
   @property({ type: Boolean })
   busy = false;
 
@@ -124,32 +118,11 @@ export class SettingsView extends LitElement {
   @property({ attribute: false })
   shellyDiscoveryMessage?: string | null;
 
-  @state()
-  private shellyIp?: string;
-
-  @state()
-  private shellyUsername?: string;
-
-  @state()
-  private shellyPassword = "";
-
-  @state()
-  private clearShellyPassword = false;
-
-  @state()
-  private kasaIp?: string;
-
   private readonly form = createRef<HTMLFormElement>();
-
-  @state()
-  private measureDeviceValue = "";
-
-  @state()
-  private hassPowerEntity = "";
 
   static readonly styles = [sharedStyles, css`
     :host { display: block; min-width: 0; max-width: 100%; }
-    measure-settings-github-section { display: contents; }
+    measure-settings-github-section, measure-settings-power-meter-section { display: contents; }
     form { display: grid; gap: 1rem; min-width: 0; max-width: 100%; margin-top: 1rem; }
     .settings-layout { display: grid; grid-template-columns: minmax(180px, 0.32fr) minmax(0, 1fr); gap: 1.25rem; align-items: start; }
     .settings-nav { display: grid; gap: 0.4rem; padding: 0.45rem; border: 1px solid var(--line); border-radius: 12px; background: var(--field); }
@@ -208,8 +181,6 @@ export class SettingsView extends LitElement {
 
   willUpdate(changedProperties: PropertyValues<this>) {
     if (changedProperties.has("settings")) {
-      this.measureDeviceValue = this.settings?.default_measure_device ?? "";
-      this.hassPowerEntity = this.settings?.default_power_entity_id ?? "";
       this.contributorGithubValue = undefined;
     }
     // Honour a requested section (e.g. opened from the GitHub contribution shortcut) once,
@@ -222,14 +193,14 @@ export class SettingsView extends LitElement {
 
   protected async getUpdateComplete(): Promise<boolean> {
     const complete = await super.getUpdateComplete();
-    const github = this.shadowRoot?.querySelector<LitElement>("measure-settings-github-section");
-    await github?.updateComplete;
+    const sections = this.shadowRoot?.querySelectorAll<LitElement>(
+      "measure-settings-github-section, measure-settings-power-meter-section",
+    );
+    await Promise.all([...sections ?? []].map((section) => section.updateComplete));
     return complete;
   }
 
   render() {
-    const powerMeter = this.meter ?? this.settings?.power_meter ?? "hass";
-    const descriptor = meterFor(powerMeter);
     const connectedGithubUsername = this.contributionAuth?.connected ? this.contributionAuth.identity?.login : "";
     const contributorGithub = this.contributorGithubValue
       ?? (this.settings?.default_contributor_github || connectedGithubUsername || "");
@@ -254,45 +225,23 @@ export class SettingsView extends LitElement {
             <section class="settings-section" ?hidden=${this.activeSection !== "power_meter"} aria-labelledby="power-meter-title">
               <h3 id="power-meter-title">Power meter</h3>
               <p class="muted">Choose where readings come from and set the hardware metadata added to new profiles.</p>
-              <div class="section-fields">
-                <measure-combobox
-                  name="default_measure_device"
-                  label="Power measurement device"
-                  .value=${this.measureDeviceValue}
-                  .options=${this.measureDeviceOptions()}
-                  placeholder="e.g. Shelly Plug S"
-                  .hint=${this.measureDevicesLoading
-                    ? "Loading names used by existing Powercalc profiles…"
-                    : "Manufacturer and model of the meter used to take readings. This is prefilled in each profile and can still be changed there."}
-                  required
-                  allowCustom
-                  @combobox-change=${this.measureDeviceChanged}
-                >
-                  <input slot="value" type="hidden" name="default_measure_device" .value=${this.measureDeviceValue} />
-                </measure-combobox>
-                ${this.measureDevicesError
-                  ? html`<small class="field-hint error" role="status">Library suggestions are unavailable; manual entry still works.</small>`
-                  : nothing}
-                <label>
-                  <span>Power measurement device firmware</span>
-                  <input
-                    name="default_measure_device_firmware"
-                    .value=${this.settings?.default_measure_device_firmware ?? ""}
-                    autocomplete="off"
-                    placeholder="Optional firmware version"
-                  />
-                  <small class="field-hint">Prefilled in new profile metadata and editable for each profile.</small>
-                </label>
-                ${optionSelect("power_meter", "Type", POWER_METER_LIST.map((meter) => ({ value: meter.type, label: meter.label })), {
-                  selected: powerMeter,
-                  required: true,
-                  placeholder: "Select a power meter type",
-                  onChange: this.powerMeterChanged,
-                })}
-                ${this.renderMeterFields(powerMeter)}
-                ${descriptor.qualityNote ? html`<p class="quality-requirements">${descriptor.qualityNote}</p>` : nothing}
-                ${descriptor.validatable ? this.renderTestRow() : nothing}
-              </div>
+              <measure-settings-power-meter-section
+                .powers=${this.powers}
+                .settings=${this.settings}
+                .measureDevices=${this.measureDevices}
+                .measureDevicesLoading=${this.measureDevicesLoading}
+                .measureDevicesError=${this.measureDevicesError}
+                .busy=${this.busy}
+                .testing=${this.testing}
+                .testResult=${this.testResult}
+                .shellyDiscoveryDevices=${this.shellyDiscoveryDevices}
+                .discoveringShellys=${this.discoveringShellys}
+                .shellyDiscoveryError=${this.shellyDiscoveryError}
+                .shellyDiscoveryAvailable=${this.shellyDiscoveryAvailable}
+                .shellyDiscoveryMessage=${this.shellyDiscoveryMessage}
+                @power-meter-test=${this.test}
+                @test-clear=${this.clearTestResult}
+              ></measure-settings-power-meter-section>
             </section>
 
             <section class="settings-section" ?hidden=${this.activeSection !== "profile"} aria-labelledby="profile-metadata-title">
@@ -374,147 +323,8 @@ export class SettingsView extends LitElement {
     `;
   }
 
-  private measureDeviceOptions(): ComboboxOption[] {
-    if (this.measureDevicesLoading || this.measureDevicesError) return [];
-    return this.measureDevices.map((device) => ({ value: device, label: device }));
-  }
-
   private contributorGithubChanged(event: Event): void {
     this.contributorGithubValue = (event.target as HTMLInputElement).value;
-  }
-
-  private measureDeviceChanged(event: CustomEvent<{ value: string }>): void {
-    this.measureDeviceValue = event.detail.value;
-  }
-
-  private hassPowerEntityChanged(event: CustomEvent<{ value: string }>): void {
-    this.hassPowerEntity = event.detail.value;
-    this.powerMeterSettingsChanged();
-  }
-
-  /**
-   * The inputs each meter needs. Kept here rather than in the registry because they are bound to
-   * this view's handlers and credential state; the record's type still names any meter left out.
-   */
-  private renderMeterFields(type: PowerMeterType) {
-    const fields: Record<PowerMeterType, () => unknown> = {
-      hass: () => this.renderHassFields(),
-      shelly: () => this.renderShellyFields(),
-      kasa: () => this.renderKasaFields(),
-      dummy: () => nothing,
-    };
-    return fields[type]();
-  }
-
-  private renderHassFields() {
-    const options = this.powers.map((entity) => ({
-      value: entity.entity_id,
-      label: `${entity.name} · ${entity.entity_id}`,
-    }));
-    return html`
-      <measure-combobox
-        name="default_power_entity_id"
-        label="Power sensor"
-        .value=${this.hassPowerEntity}
-        .options=${options}
-        placeholder="Search power sensors"
-        required
-        @combobox-change=${this.hassPowerEntityChanged}
-      >
-        <input slot="value" type="hidden" name="default_power_entity_id" .value=${this.hassPowerEntity} />
-      </measure-combobox>`;
-  }
-
-  private renderTestRow() {
-    return html`
-      <div class="test-row">
-        <button type="button" @click=${this.test} ?disabled=${this.testing || this.busy}>${this.testing ? "Validating…" : "Validate measurement device"}</button>
-        ${this.renderTestResult()}
-      </div>`;
-  }
-
-  private renderTestResult() {
-    if (!this.testResult) return nothing;
-    return html`<measure-power-meter-diagnostic .diagnostic=${this.testResult}></measure-power-meter-diagnostic>`;
-  }
-
-  private renderShellyFields() {
-    const address = this.shellyIp ?? this.settings?.shelly_ip ?? "";
-    return html`
-      <div class="discovery">
-        <div class="discovery-header">
-          <strong>Discovered Shelly devices</strong>
-          <button type="button" @click=${this.discoverShellys} ?disabled=${this.discoveringShellys || this.busy}>
-            ${this.discoveringShellys ? "Searching…" : "Refresh"}
-          </button>
-        </div>
-        ${this.renderShellyDiscovery(address)}
-      </div>
-      <label>
-        <span>Shelly IP address</span>
-        <input name="shelly_ip" .value=${address} required autocomplete="off" placeholder="192.168.1.50" @input=${this.shellyIpChanged} />
-        <small class="field-hint">Select a discovered device above or enter its IP address manually.</small>
-      </label>
-      <div class="grid">
-        <label>
-          <span>Shelly username</span>
-          <input name="shelly_username" .value=${this.shellyUsername ?? this.settings?.shelly_username ?? DEFAULT_SHELLY_USERNAME} required autocomplete="username" maxlength="50" @input=${this.shellyUsernameChanged} />
-          <small class="field-hint">Gen1 devices may use a custom username. Gen2 and newer always use admin.</small>
-        </label>
-        <label>
-          <span>Shelly password</span>
-          <input name="shelly_password" type="password" .value=${this.shellyPassword} autocomplete="new-password" maxlength="255" placeholder=${this.settings?.shelly_password_configured ? "Saved password (leave blank to keep)" : "Optional"} @input=${this.shellyPasswordChanged} />
-          <small class="field-hint">Stored privately in the app and never returned by the API.</small>
-        </label>
-      </div>
-      ${this.renderClearShellyPassword()}`;
-  }
-
-  private renderClearShellyPassword() {
-    if (!this.settings?.shelly_password_configured) return nothing;
-    return html`
-      <label class="check">
-        <input name="clear_shelly_password" type="checkbox" .checked=${this.clearShellyPassword} @change=${this.clearShellyPasswordChanged} />
-        <span>Remove the saved Shelly password</span>
-      </label>`;
-  }
-
-  private renderShellyDiscovery(selectedAddress: string) {
-    if (this.discoveringShellys) return html`<p class="discovery-status" role="status">Searching for Shelly devices on your network…</p>`;
-    if (this.shellyDiscoveryError) return html`<p class="discovery-status error" role="alert">${this.shellyDiscoveryError}</p>`;
-    if (this.shellyDiscoveryAvailable === false) {
-      return html`<p class="discovery-status">${this.shellyDiscoveryMessage ?? "Shelly discovery is unavailable. Enter the IP address manually."}</p>`;
-    }
-    if (!this.shellyDiscoveryDevices.length) return html`<p class="discovery-status">No Shelly devices found. You can refresh or enter an IP address manually.</p>`;
-    return optionSelect("discovered_shelly", "Select device", [
-      { value: "", label: "Select a discovered Shelly" },
-      ...this.shellyDiscoveryDevices.map((device) => ({
-        value: device.ip_address,
-        label: this.shellyDeviceLabel(device),
-        disabled: !device.supported && !device.auth_required,
-      })),
-    ], {
-      selected: selectedAddress,
-      placeholder: "Search discovered Shelly devices",
-      onChange: this.discoveredShellyChanged,
-    });
-  }
-
-  private renderKasaFields() {
-    const address = this.kasaIp ?? this.settings?.kasa_ip ?? "";
-    return html`
-      <label>
-        <span>Kasa IP address</span>
-        <input name="kasa_ip" .value=${address} required autocomplete="off" placeholder="192.168.1.50" @input=${this.kasaIpChanged} />
-        <small class="field-hint">Enter the IP address of a Kasa plug with energy monitoring, such as a KP115 or HS110. Give it a static lease in your router so it stays reachable.</small>
-      </label>`;
-  }
-
-  private shellyDeviceLabel(device: ShellyDiscoveryDevice): string {
-    const identity = [device.name, device.model, device.generation === null ? null : `Gen ${device.generation}`, device.ip_address]
-      .filter((part): part is string => Boolean(part))
-      .join(" · ");
-    return device.supported ? identity : `${identity} — ${device.reason ?? "Not supported"}`;
   }
 
   private collect(): AppSettingsUpdate | null {
@@ -573,59 +383,8 @@ export class SettingsView extends LitElement {
     emit<AppSettingsUpdate>(this, "test", settings);
   }
 
-  private powerMeterChanged(event: Event): void {
-    this.clearTestResult();
-    // Keep the choice in local state so an app-shell re-render can't clobber the
-    // in-progress form (which would reset the meter type and typed device IP).
-    this.meter = (event.currentTarget as HTMLInputElement).value as PowerMeterType;
-    if (meterFor(this.meter).discoverable) this.discoverShellys();
-  }
-
-  private powerMeterSettingsChanged(): void {
-    this.clearTestResult();
-  }
-
-  private shellyIpChanged(event: Event): void {
-    this.shellyIp = (event.currentTarget as HTMLInputElement).value;
-    this.powerMeterSettingsChanged();
-  }
-
-  private shellyUsernameChanged(event: Event): void {
-    this.shellyUsername = (event.currentTarget as HTMLInputElement).value;
-    this.powerMeterSettingsChanged();
-  }
-
-  private shellyPasswordChanged(event: Event): void {
-    this.shellyPassword = (event.currentTarget as HTMLInputElement).value;
-    this.clearShellyPassword = false;
-    this.powerMeterSettingsChanged();
-  }
-
-  private clearShellyPasswordChanged(event: Event): void {
-    this.clearShellyPassword = (event.currentTarget as HTMLInputElement).checked;
-    if (this.clearShellyPassword) this.shellyPassword = "";
-    this.powerMeterSettingsChanged();
-  }
-
-  private kasaIpChanged(event: Event): void {
-    this.kasaIp = (event.currentTarget as HTMLInputElement).value;
-    this.powerMeterSettingsChanged();
-  }
-
-  private discoveredShellyChanged(event: Event): void {
-    const address = (event.currentTarget as HTMLInputElement).value;
-    if (!address) return;
-    this.shellyIp = address;
-    this.powerMeterSettingsChanged();
-  }
-
-  private discoverShellys(): void {
-    emit(this, "shelly-discover");
-  }
-
   private clearTestResult(): void {
     this.testResult = undefined;
-    emit(this, "test-clear");
   }
 
   private selectSection(section: SettingsSection): void {
